# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Trip Builder API",
        "description": "Manage flights and trips ",
        "contact": {
            "email": "liamlian@mail.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost/fulltrip/public",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
        "/airports": {
            "get": {
                "tags": [
                    "Trip"
                ],
                "summary": "Get list of airports",
                "description": "Get an alphabetical list of real world airports",
                "operationId": "airports",
                "parameters": [
                    {
                        "name": "order",
                        "in": "query",
                        "description": "order by, name or code (default to be code)",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/trips": {
            "get": {
                "tags": [
                    "Trip"
                ],
                "summary": "Get list of trips",
                "description": "Get the list of current trips",
                "operationId": "trips",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/trip/{id}": {
            "get": {
                "tags": [
                    "Trip"
                ],
                "summary": "Get info of a specific trip",
                "description": "Get all details of a specific trip.",
                "operationId": "trip",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "trip id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/trip/{id}/flights": {
            "get": {
                "tags": [
                    "Trip"
                ],
                "summary": "Get list of flights under a trip",
                "description": "Get the flights under a trip.",
                "operationId": "tripFlights",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "trip id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/trip/{id}/add": {
            "post": {
                "tags": [
                    "Trip"
                ],
                "summary": "Add Flight",
                "description": "Add a new flight into a trip",
                "operationId": "addFlight",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "the id of the trip to be added to",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "from",
                        "in": "header",
                        "description": "leaving from",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "to",
                        "in": "header",
                        "description": "going to",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/trip/{tripId}/{refId}": {
            "delete": {
                "tags": [
                    "Trip"
                ],
                "summary": "Remove Flight",
                "description": "Remove a flight from a trip",
                "operationId": "removeFlight",
                "parameters": [
                    {
                        "name": "tripId",
                        "in": "path",
                        "description": "the id of the trip to be removed from",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "refId",
                        "in": "path",
                        "description": "the unique reference id of the flight",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        }
    },
    "definitions": {}
}